# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SDN/AAC Internal Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /training-schedules:
    get:
      operationId: listTrainingSchedules
      tags: [trainingSchedules]
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Training schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingSchedule"
    post:
      operationId: createTrainingSchedule
      tags: [trainingSchedules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSchedule"
  /training-schedules/{id}:
    patch:
      operationId: updateTrainingSchedule
      tags: [trainingSchedules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSchedule"
    delete:
      operationId: deleteTrainingSchedule
      tags: [trainingSchedules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Training Sessions ─────────────────────────────────────────────────────────
  /training-sessions:
    get:
      operationId: listTrainingSessions
      tags: [trainingSessions]
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
        - name: groupCategory
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: trainerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Training sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingSession"
    post:
      operationId: createTrainingSession
      tags: [trainingSessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSession"
...
  /training-sessions/{id}:
    get:
      operationId: getTrainingSession
      tags: [trainingSessions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Training session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSession"
    patch:
      operationId: updateTrainingSession
      tags: [trainingSessions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingSessionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingSession"
    delete:
      operationId: deleteTrainingSession
      tags: [trainingSessions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Attendance ────────────────────────────────────────────────────────────────
  /training-sessions/{id}/attendance:
    get:
      operationId: getSessionAttendance
      tags: [attendance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    put:
      operationId: saveSessionAttendance
      tags: [attendance]
      summary: Save all attendance records for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceSaveInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /athletes/{id}/attendance-summary:
    get:
      operationId: getAthleteAttendanceSummary
      tags: [attendance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"

  # ── Competitions ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TrainingSchedule:
      type: object
      required: [id, seasonId, groupCategory, daysOfWeek, startTime, endTime, trainingType]
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        groupCategory:
          type: string
        daysOfWeek:
          type: array
          items:
            type: integer
        startTime:
          type: string
        endTime:
          type: string
        trainingType:
          type: string
          enum: [agua, ginasio, ergometro, outro]
        trainerIds:
          type: array
          items:
            type: integer
        notes:
          type: ["string", "null"]
    TrainingScheduleInput:
      type: object
      required: [seasonId, groupCategory, daysOfWeek, startTime, endTime, trainingType]
      properties:
        seasonId:
          type: integer
        groupCategory:
          type: string
        daysOfWeek:
          type: array
          items:
            type: integer
        startTime:
          type: string
        endTime:
          type: string
        trainingType:
          type: string
          enum: [agua, ginasio, ergometro, outro]
        trainerIds:
          type: array
          items:
            type: integer
        notes:
          type: ["string", "null"]
    TrainingScheduleUpdate:
      type: object
      properties:
        daysOfWeek:
          type: array
          items:
            type: integer
        startTime:
          type: string
        endTime:
          type: string
        trainingType:
          type: string
          enum: [agua, ginasio, ergometro, outro]
        trainerIds:
          type: array
          items:
            type: integer
        notes:
          type: ["string", "null"]

    # ── Training Sessions ───────────────────────────────────────────────────────
    TrainingSession:
      type: object
      required: [id, date, groupCategory, trainingType, startTime, endTime]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        groupCategory:
          type: string
        trainingType:
          type: string
          enum: [agua, ginasio, ergometro, prova, estagio, outro]
        startTime:
          type: string
        endTime:
          type: string
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        seasonId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        attendanceCount:
          type: ["integer", "null"]
...
    TrainingSessionInput:
      type: object
      required: [date, groupCategory, trainingType, startTime, endTime]
      properties:
        date:
          type: string
          format: date
        groupCategory:
          type: string
        trainingType:
          type: string
          enum: [agua, ginasio, ergometro, prova, estagio, outro]
        startTime:
          type: string
        endTime:
          type: string
        trainerId:
          type: ["integer", "null"]
        seasonId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    TrainingSessionUpdate:
      type: object
      properties:
        trainingType:
          type: string
          enum: [agua, ginasio, ergometro, prova, estagio, outro]
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: ["string", "null"]

    # ── Attendance ──────────────────────────────────────────────────────────────
    AttendanceRecord:
      type: object
      required: [id, sessionId, athleteId, status]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        athleteId:
          type: integer
        athleteName:
          type: string
        status:
          type: string
          enum: [presente, ausente, ausente_justificado]
        observation:
          type: ["string", "null"]
    AttendanceSaveInput:
      type: object
      required: [records]
      properties:
        records:
          type: array
          items:
            type: object
            required: [athleteId, status]
            properties:
              athleteId:
                type: integer
              status:
                type: string
                enum: [presente, ausente, ausente_justificado]
              observation:
                type: ["string", "null"]
    AttendanceSummary:
      type: object
      required: [athleteId, athleteName, totalSessions, present, absent, absentJustified, attendanceRate]
      properties:
        athleteId:
          type: integer
        athleteName:
          type: string
        totalSessions:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        absentJustified:
          type: integer
        attendanceRate:
          type: number
        byType:
          type: object
          additionalProperties:
            type: integer

    # ── Competitions ────────────────────────────────────────────────────────────
...